+++ src/aks-preview/azext_aks_preview/agentpool_decorator.py
@@ -51,6 +51,7 @@ from azext_aks_preview._consts import (
     CONST_GPU_DRIVER_NONE,
     CONST_GPU_MANAGEMENT_MODE_MANAGED,
     CONST_GPU_MANAGEMENT_MODE_UNMANAGED,
+    CONST_MANAGED_GPU_DRIVER_MODE_DEVICE_PLUGIN,
     CONST_NODEPOOL_MODE_MANAGEDSYSTEM,
     CONST_NODEPOOL_MODE_MACHINES,
     CONST_OS_SKU_WINDOWS2025,
@@ -334,6 +335,10 @@ class AKSPreviewAgentPoolContext(AKSAgentPoolContext):
                 protocol=r[0][5].upper(),
             ))
         return port_ranges
+
+    def get_enable_managed_dranet(self) -> bool:
+        """Obtain the value of enable_managed_dranet."""
+        return self.raw_param.get("enable_managed_dranet")
 
     def get_ip_tags(self) -> Union[List[IPTag], None]:
         ip_tags = self.raw_param.get("node_public_ip_tags")
@@ -670,6 +675,21 @@ class AKSPreviewAgentPoolContext(AKSAgentPoolContext):
                     self.agentpool.gpu_profile.nvidia.management_mode == CONST_GPU_MANAGEMENT_MODE_MANAGED
                 )
         return enable_managed_gpu
+
+    def get_managed_gpu_driver_mode(self) -> Union[str, None]:
+        """Obtain the value of managed_gpu_driver_mode."""
+        managed_gpu_driver_mode = self.raw_param.get("managed_gpu_driver_mode")
+        enable_managed_gpu = self.get_enable_managed_gpu()
+
+        if managed_gpu_driver_mode is not None and enable_managed_gpu is not True:
+            raise ArgumentUsageError(
+                "--managed-gpu-driver-mode requires --enable-managed-gpu to be set to true."
+            )
+
+        if managed_gpu_driver_mode is None and enable_managed_gpu is True:
+            managed_gpu_driver_mode = CONST_MANAGED_GPU_DRIVER_MODE_DEVICE_PLUGIN
+
+        return managed_gpu_driver_mode
 
     def get_disable_artifact_streaming(self) -> bool:
         """Obtain the value of disable_artifact_streaming.
@@ -1438,6 +1458,11 @@ class AKSPreviewAgentPoolAddDecorator(AKSAgentPoolAddDecorator):
         if secondary_nics is not None:
             agentpool.network_profile.secondary_network_interfaces = secondary_nics
 
+        if self.context.get_enable_managed_dranet():
+            agentpool.network_profile.dranet = self.models.DRANETProfile(
+                mode="Managed"
+            )
+
         return agentpool
 
     def set_up_taints(self, agentpool: AgentPool) -> AgentPool:
@@ -1487,6 +1512,7 @@ class AKSPreviewAgentPoolAddDecorator(AKSAgentPoolAddDecorator):
         self._ensure_agentpool(agentpool)
 
         enable_managed_gpu = self.context.get_enable_managed_gpu()
+        managed_gpu_driver_mode = self.context.get_managed_gpu_driver_mode()
 
         if enable_managed_gpu:
             if agentpool.gpu_profile is None:
@@ -1494,6 +1520,7 @@ class AKSPreviewAgentPoolAddDecorator(AKSAgentPoolAddDecorator):
             if agentpool.gpu_profile.nvidia is None:
                 agentpool.gpu_profile.nvidia = self.models.NvidiaGPUProfile()  # pylint: disable=no-member
             agentpool.gpu_profile.nvidia.management_mode = CONST_GPU_MANAGEMENT_MODE_MANAGED
+            agentpool.gpu_profile.nvidia.driver_mode = managed_gpu_driver_mode
             agentpool.gpu_profile.driver = CONST_GPU_DRIVER_INSTALL
         return agentpool
 
@@ -1941,12 +1968,17 @@ class AKSPreviewAgentPoolUpdateDecorator(AKSAgentPoolUpdateDecorator):
 
         asg_ids = self.context.get_asg_ids()
         allowed_host_ports = self.context.get_allowed_host_ports()
-        if not agentpool.network_profile and (asg_ids or allowed_host_ports):
+        enable_managed_dranet = self.context.get_enable_managed_dranet()
+        if not agentpool.network_profile and (asg_ids is not None or allowed_host_ports is not None or enable_managed_dranet):
             agentpool.network_profile = self.models.AgentPoolNetworkProfile()  # pylint: disable=no-member
         if asg_ids is not None:
             agentpool.network_profile.application_security_groups = asg_ids
         if allowed_host_ports is not None:
             agentpool.network_profile.allowed_host_ports = allowed_host_ports
+        if enable_managed_dranet:
+            agentpool.network_profile.dranet = self.models.DRANETProfile(
+                mode="Managed"
+            )
         return agentpool
 
     def update_gpu_profile(self, agentpool: AgentPool) -> AgentPool:
@@ -1998,6 +2030,7 @@ class AKSPreviewAgentPoolUpdateDecorator(AKSAgentPoolUpdateDecorator):
         self._ensure_agentpool(agentpool)
 
         enable_managed_gpu = self.context.get_enable_managed_gpu()
+        managed_gpu_driver_mode = self.context.get_managed_gpu_driver_mode()
         if enable_managed_gpu is None:
             return agentpool
 
@@ -2007,6 +2040,7 @@ class AKSPreviewAgentPoolUpdateDecorator(AKSAgentPoolUpdateDecorator):
             if agentpool.gpu_profile.nvidia is None:
                 agentpool.gpu_profile.nvidia = self.models.NvidiaGPUProfile()  # pylint: disable=no-member
             agentpool.gpu_profile.nvidia.management_mode = CONST_GPU_MANAGEMENT_MODE_MANAGED
+            agentpool.gpu_profile.nvidia.driver_mode = managed_gpu_driver_mode
             agentpool.gpu_profile.driver = CONST_GPU_DRIVER_INSTALL
         else:
             if agentpool.gpu_profile and agentpool.gpu_profile.nvidia: