+++ src/quantum/azext_quantum/commands.py
@@ -90,6 +90,19 @@ def transform_offerings(offerings):
     return [one(offering) for offering in offerings]
 
 
+def transform_suite_offers(suite_offers):
+    def one(offer):
+        properties = offer['properties']
+        return OrderedDict([
+            ('Provider ID', properties['providerId']),
+            ('Provider Name', properties['providerName']),
+            ('Company', properties['companyName']),
+            ('Location', properties['location'])
+        ])
+
+    return [one(offer) for offer in suite_offers]
+
+
 def transform_output(results):
     def one(key, value):
         repeat = round(20 * value)
@@ -145,6 +158,7 @@ def load_command_table(self, _):
     job_ops = CliCommandType(operations_tmpl='azext_quantum.operations.job#{}')
     target_ops = CliCommandType(operations_tmpl='azext_quantum.operations.target#{}')
     offerings_ops = CliCommandType(operations_tmpl='azext_quantum.operations.offerings#{}')
+    suite_offers_ops = CliCommandType(operations_tmpl='azext_quantum.operations.suite_offers#{}')
 
     with self.command_group('quantum workspace', workspace_ops) as w:
         w.command('create', 'create')
@@ -189,3 +203,6 @@ def load_command_table(self, _):
         o.command('list', 'list_offerings', table_transformer=transform_offerings)
         o.command('accept-terms', 'accept_terms', validator=validate_provider_and_sku_info)
         o.command('show-terms', 'show_terms', validator=validate_provider_and_sku_info)
+
+    with self.command_group('quantum suite-offer', suite_offers_ops) as s:
+        s.command('list', 'list_suite_offers', table_transformer=transform_suite_offers)